Added a custom cmc component and fixed all replication errors
Replicating  AddMovementInput is not required, and I made it a default for it to pass the MaxWalkSpeed into the network call in as it's just as important as acceleration and is changed every frame. Other than that I cleaned up some of the code, and the only discrepancies between the client and server have to deal with the onMovementMode changed values. That's where we'll apply rampsliding and trimping in the cmc component itself, and fix whatever's causing the errors in the old code